% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.lotri.R
\name{as.lotri}
\alias{as.lotri}
\alias{as.lotri.matrix}
\alias{as.lotri.data.frame}
\alias{as.lotri.default}
\title{As lower triangular matrix}
\usage{
as.lotri(x, ..., default = "")

\method{as.lotri}{matrix}(x, ..., default = "")

\method{as.lotri}{data.frame}(x, ..., default = "")

\method{as.lotri}{default}(x, ..., default = "")
}
\arguments{
\item{x}{Matrix or other data frame}

\item{...}{Other factors}

\item{default}{Is the default factor when no conditioning is
implemented.}
}
\value{
Lower triangular matrix
}
\description{
As lower triangular matrix
}
\author{
Matthew Fidler
}
