% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{asLotriMat}
\alias{asLotriMat}
\title{Convert Matrix to lotri Object}
\usage{
asLotriMat(x, extra, def)
}
\arguments{
\item{x}{A completely named symmetric matrix}

\item{extra}{A list of extra properties (lower, upper bounds, etc.)}

\item{def}{Default name for the matrix block (string of length 1)}
}
\value{
A lotri object
}
\description{
Convert a named symmetric matrix to a lotri object with optional
  bounds and other properties.
}
\keyword{internal}
