% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iov.R
\name{.lotriOmegaBuiltinSize}
\alias{.lotriOmegaBuiltinSize}
\title{Get the number of built-in omega inverses}
\usage{
.lotriOmegaBuiltinSize()
}
\value{
The dimension of the pre-compiled omega matrix translations
}
\description{
Get the number of built-in omega inverses
}
\examples{
.nlmixr2omegaBuiltinSize()
}
\author{
Matthew L. Fidler
}
