% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotriPtrs.R
\name{.lotriPointers}
\alias{.lotriPointers}
\title{This function is used to get the pointers to the C objects that are used in the lotri package.}
\usage{
.lotriPointers()
}
\value{
A list of function pointers
}
\description{
This function is used to get the pointers to the C objects that are used in the lotri package.
}
\examples{

.lotriPointers()
}
\author{
Matthew L. Fidler
}
\keyword{internal}
