% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{getLotriPointers}
\alias{getLotriPointers}
\title{Get lotri C Function Pointers}
\usage{
getLotriPointers()
}
\value{
A named list of external pointers to C functions
}
\description{
Get external pointers to lotri's C functions for use by other
packages.
}
\keyword{internal}
