% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{isLotri}
\alias{isLotri}
\title{Check if Object is a lotri Object}
\usage{
isLotri(lotri)
}
\arguments{
\item{lotri}{An R object to check}
}
\value{
Logical indicating if the object is a lotri object
}
\description{
Check if an R object is a valid lotri object.
}
\keyword{internal}
