% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{lotriAllNames}
\alias{lotriAllNames}
\title{Get All Names from lotri Object}
\usage{
lotriAllNames(lotri)
}
\arguments{
\item{lotri}{A lotri object (matrix or list of matrices)}
}
\value{
Character vector of all parameter names
}
\description{
Extract all parameter names from a lotri matrix object.
}
\keyword{internal}
