% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{lotriLstToMat}
\alias{lotriLstToMat}
\title{Convert List to Matrix}
\usage{
lotriLstToMat(lst_, format, startNum, matCls)
}
\arguments{
\item{lst_}{A list of matrices}

\item{format}{Format string for naming}

\item{startNum}{Starting number for formatted names}

\item{matCls}{Matrix class}
}
\value{
A block-diagonal matrix
}
\description{
Convert a list of matrices to a single block-diagonal matrix.
}
\keyword{internal}
