% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{lotriMaxNu}
\alias{lotriMaxNu}
\title{Maximum nu from lotri Object}
\usage{
lotriMaxNu(lotri)
}
\arguments{
\item{lotri}{A lotri object}
}
\value{
Maximum nu value as a double
}
\description{
Get the maximum nu value from a lotri matrix object.
}
\keyword{internal}
