% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iov.R
\name{lotriOmega}
\alias{lotriOmega}
\title{Create a nlmixr2omega for use in focei}
\usage{
lotriOmega(mat, diag.xform = c("sqrt", "log", "identity"))
}
\arguments{
\item{mat}{Matrix for conversion}

\item{diag.xform}{Form of the diagonal transformation}
}
\value{
nlmixr2omega object
}
\description{
Create a nlmixr2omega for use in focei
}
\examples{

d <- 10
m <- matrix(rnorm(d^2), d, d)
mcov <- tcrossprod(m, m)
ome <-nlmixr2omega(mcov)

}
\author{
Matthew L. Fidler
}
