% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{nearPD_}
\alias{nearPD_}
\title{Nearest Positive Definite Matrix}
\usage{
nearPD_(
  x_r,
  keepDiag,
  do2eigen,
  doDykstra,
  only_values,
  eig_tol,
  conv_tol,
  posd_tol,
  maxit,
  trace
)
}
\arguments{
\item{x_r}{A symmetric matrix}

\item{keepDiag}{Logical indicating if the diagonal should be preserved}

\item{do2eigen}{Logical indicating if eigenvalue step should be performed}

\item{doDykstra}{Logical indicating if Dykstra's correction should be used}

\item{only_values}{Logical indicating if only eigenvalues should be returned}

\item{eig_tol}{Tolerance for eigenvalue positiveness}

\item{conv_tol}{Convergence tolerance for the algorithm}

\item{posd_tol}{Tolerance for enforcing positive definiteness}

\item{maxit}{Maximum number of iterations}

\item{trace}{Logical indicating if iterations should be traced}
}
\value{
The nearest positive definite matrix, or eigenvalues if only_values is
TRUE
}
\description{
Compute the nearest positive definite matrix to an approximate one,
  typically a correlation or variance-covariance matrix.
}
\keyword{internal}
