% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{omegaBlock}
\alias{omegaBlock}
\title{Get the omega block theta information}
\usage{
omegaBlock(omega, diagXform = c("sqrt", "log", "identity"))
}
\arguments{
\item{omega}{this is the omega matrix}

\item{diagXform}{diagonal transform}
}
\description{
Get the omega block theta information
}
\examples{

d <- 10

m <- matrix(rnorm(d^2), d, d)

mcov <- tcrossprod(m, m)

omegaBlock(mcov)

}
\author{
Matthew L. Fidler
}
