% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotri.R
\name{lotriMat}
\alias{lotriMat}
\title{Create a matrix from a list of matrices}
\usage{
lotriMat(list)
}
\arguments{
\item{list}{List of symmetric named matrices}
}
\value{
Named symmetric block diagonal matrix based on
  concatenating the list of matrices together
}
\description{
Create a matrix from a list of matrices
}
